add: local controller
added local usage for the API instead of using aws-sdk.
using sqlite as database and two tables, bucket and object.
table bucket is used to store references for the bucket (a tmp folder
with the name of the bucket is created to store the files), table
object is used to store references for the objects under a bucket.